Added the handlers for verifying account requests.
When an account is requested a message is sent to the email of the user
requesting the account. This message contains a link with a verification
code that is needed to create the account.
When that link is visited the newly created handlers compare the code
with the one stored on the server and the user is free to create an
account if they match.